+++ spa/src/api/deferral/index.ts
@@ -3,12 +3,12 @@ import {
 	DeferredInstallationUrlParams,
 	GetDeferredInstallationUrl
 } from "rest-interfaces";
-import { axiosRest, axiosRestWithNoJwt, axiosRestWithNoJwtButWithGitHubToken } from "../axiosInstance";
+import { axiosRest, axiosRestWithNoJwtButWithGitHubToken } from "../axiosInstance";
 
 export default {
-	parseDeferredRequestId: (requestId: string) => axiosRestWithNoJwt.get<DeferralParsedRequest>(`/rest/app/cloud/deferred/parse/${requestId}`),
 	getDeferredInstallationUrl: (params: DeferredInstallationUrlParams) =>
 		axiosRest.get<GetDeferredInstallationUrl>("/rest/app/cloud/deferred/installation-url", { params }),
-	connectDeferredOrg: (requestId: string) => axiosRestWithNoJwtButWithGitHubToken.post(`/rest/app/cloud/deferred/connect/${requestId}`)
+	connectDeferredOrg: (requestId: string) => axiosRestWithNoJwtButWithGitHubToken.post(`/rest/app/cloud/deferred/connect/${requestId}`),
+	parseDeferredRequestId: (requestId: string) => axiosRestWithNoJwtButWithGitHubToken.get<DeferralParsedRequest>(`/rest/app/cloud/deferred/parse/${requestId}`)
 };
